cast lda to size_t in calculation of stride
[ROCm/hipBLAS commit: 75ace65a5cc6fd2f3383aae5712d7e35c8309997]

+++ clients/include/testing_syrkx_strided_batched.hpp
@@ -49,9 +49,9 @@ hipblasStatus_t testing_syrkx_strided_batched(const Arguments& argus)
     }
 
     int           K1       = transA == HIPBLAS_OP_N ? K : N;
-    hipblasStride stride_A = lda * K1 * stride_scale;
-    hipblasStride stride_B = ldb * K1 * stride_scale;
-    hipblasStride stride_C = ldc * N * stride_scale;
+    hipblasStride stride_A = size_t(lda) * K1 * stride_scale;
+    hipblasStride stride_B = size_t(ldb) * K1 * stride_scale;
+    hipblasStride stride_C = size_t(ldc) * N * stride_scale;
     int           A_size   = stride_A * batch_count;
     int           B_size   = stride_B * batch_count;
     int           C_size   = stride_C * batch_count;
